appointment [Checkbox backend]

+++ function/toapt-cancel.php
@@ -17,11 +17,13 @@
         if(mysqli_query($connect,$delete))
 		{
             $filename = $qrfile;
-
-            if (unlink($filename)) {
-                echo 'The file ' . $filename . ' was deleted successfully!';
-            } else {
-                echo 'There was a error deleting the file ' . $filename;
+            if($filename != NULL|| ""){
+                if (unlink($filename)) {
+                    echo 'The file ' . $filename . ' was deleted successfully!';
+                }
+                else {
+                    echo 'There was a error deleting the file ' . $filename;
+                }
             }
 			echo "<script>alert('Cancel Success')</script>";	
             header("refresh: 0; url= ../index.php");
